+++ includes/time_evolution.hpp
@@ -6,28 +6,28 @@ using namespace std;
 
 class time_evolution{
 	private:
-		float lambda, eta, woof;
+		double lambda, eta, woof;
 		int kreiss;
 	public:
 		time_evolution();
-		float get_lambda();
-		float get_eta();
-		float the_Potential(float field);
-		float dPotentialdfield(float field);
-		std::vector<float> rk1_phi(std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> rk1_psi(std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> rk2_phi(std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> rk2_psi(std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> rk4_phi(std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> rk4_psi(std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> l0_phi(std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> l0_psi(std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> l1_phi(std::vector<float> i_l0, std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> l1_psi(std::vector<float> i_l0, std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> l2_phi(std::vector<float> i_l1, std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> l2_psi(std::vector<float> i_l1, std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> l3_phi(std::vector<float> i_l2, std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
-		std::vector<float> l3_psi(std::vector<float> i_l2, std::vector<float> i_vector, std::vector<float> ii_vector, float i_dx, float i_alpha);
+		double get_lambda();
+		double get_eta();
+		double the_Potential(double field);
+		double dPotentialdfield(double field);
+		std::vector<double> rk1_phi(std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> rk1_psi(std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> rk2_phi(std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> rk2_psi(std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> rk4_phi(std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> rk4_psi(std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> l0_phi(std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> l0_psi(std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> l1_phi(std::vector<double> i_l0, std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> l1_psi(std::vector<double> i_l0, std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> l2_phi(std::vector<double> i_l1, std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> l2_psi(std::vector<double> i_l1, std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> l3_phi(std::vector<double> i_l2, std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
+		std::vector<double> l3_psi(std::vector<double> i_l2, std::vector<double> i_vector, std::vector<double> ii_vector, double i_dx, double i_alpha);
 };
 
 #endif
